- Ids cambiados de int a long - Cambios en varios métodos para adaptarlos a long - Feed de comentarios de usuarios que sigues en inicio - CSSs unidos en uno

+++ src/main/java/es/geeko/model/Transaccion.java
@@ -22,10 +22,10 @@ public class Transaccion {
     private long id;
 
     @Column(name="idCliente", nullable = false)
-    private int idCliente;
+    private long idCliente;
 
     @Column(name = "idVendedor", nullable = false)
-    private int idVendedor;
+    private long idVendedor;
 
     @Column(name = "fecha", nullable = false)
     private Date Fecha = new java.sql.Date(Calendar.getInstance().getTime().getTime());
