+++ src/main/java/es/geeko/repository/MensajeRepository.java
@@ -3,5 +3,5 @@ package es.geeko.repository;
 import es.geeko.model.Mensaje;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MensajeRepository extends JpaRepository<Mensaje,Integer> {
+public interface MensajeRepository extends JpaRepository<Mensaje,Long> {
 }
