+++ src/main/java/es/geeko/repository/ReporteRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ReporteRepository extends JpaRepository<Reporte,Integer> {
+public interface ReporteRepository extends JpaRepository<Reporte,Long> {
 
     List <Reporte> findReportesByComentariosIsNotNull();
