+++ src/main/java/es/geeko/repository/UsuarioRepository.java
@@ -6,11 +6,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface UsuarioRepository extends JpaRepository<Usuario,Integer>{
+public interface UsuarioRepository extends JpaRepository<Usuario,Long>{
 
     Optional<Usuario> findUsuarioByEmilio(String emilio);
-    Usuario getUsuarioByIdIs(int id);
-    List<Usuario> findUsuariosByIdIs(int id);
+    Usuario getUsuarioByIdIs(long id);
+    List<Usuario> findUsuariosByIdIs(long id);
 
     //Usuarios reportados
     List<Usuario> findUsuariosByReportadoIsAndActivoIs(int reportado, int activo);
